requirements.txt
aiogram~=3.6.0
python-dotenv~=1.0.1
tortoise-orm~=0.20.1